requirements.txt
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.54
PyYAML>=6.0.0
scikit-learn>=1.5.0
statsmodels>=0.14.0
lightgbm>=4.5.0
matplotlib>=3.8.0
torch>=2.5.0
transformers>=4.45.0
accelerate>=1.0.0
huggingface_hub>=0.26.0
